+++ Microsoft.Kiota.Abstractions.Tests/MultipartBodyTests.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Kiota.Abstractions.Serialization;
+﻿using System.IO;
+using Microsoft.Kiota.Abstractions.Serialization;
 using Moq;
 using Xunit;
 
@@ -10,10 +11,22 @@ public class MultipartBodyTests
     public void KeepsFilename()
     {
         var writerMock = new Mock<ISerializationWriter>();
+        var jsonWriterMock = new Mock<ISerializationWriter>();
         var requestAdapterMock = new Mock<IRequestAdapter>();
         var serializationFactoryMock = new Mock<ISerializationWriterFactory>();
 
         var body = new MultipartBody();
+        jsonWriterMock.Setup(w => w.WriteStringValue("", "fileContent"));
+        using var ms = new MemoryStream();
+        using var sr = new StreamWriter(ms);
+
+        sr.Write("fileContent");
+        sr.Flush();
+        jsonWriterMock.Setup(w => w.GetSerializedContent()).Returns(ms);
+
+        serializationFactoryMock
+            .Setup(r => r.GetSerializationWriter("application/json"))
+            .Returns(jsonWriterMock.Object);
 
         requestAdapterMock
             .Setup(r => r.SerializationWriterFactory)
@@ -24,7 +37,7 @@ public class MultipartBodyTests
         writerMock.Setup(w => w.WriteStringValue("", "--" + body.Boundary));
         writerMock.Setup(w => w.WriteStringValue("Content-Type", "application/json"));
         writerMock.Setup(w => w.WriteStringValue("Content-Disposition", "form-data; name=\"file\"; filename=\"file.json\""));
-        writerMock.Setup(w => w.WriteStringValue("", "fileContent"));
+        writerMock.Setup(w => w.WriteByteArrayValue("", ms.ToArray()));
         writerMock.Setup(w => w.WriteStringValue("", ""));
         writerMock.Setup(w => w.WriteStringValue("", "--" + body.Boundary + "--"));
 
@@ -40,10 +53,22 @@ public class MultipartBodyTests
     public void WorksWithoutFilename()
     {
         var writerMock = new Mock<ISerializationWriter>();
+        var jsonWriterMock = new Mock<ISerializationWriter>();
         var requestAdapterMock = new Mock<IRequestAdapter>();
         var serializationFactoryMock = new Mock<ISerializationWriterFactory>();
 
         var body = new MultipartBody();
+        jsonWriterMock.Setup(w => w.WriteStringValue("", "fileContent"));
+        using var ms = new MemoryStream();
+        using var sr = new StreamWriter(ms);
+
+        sr.Write("fileContent");
+        sr.Flush();
+        jsonWriterMock.Setup(w => w.GetSerializedContent()).Returns(ms);
+
+        serializationFactoryMock
+            .Setup(r => r.GetSerializationWriter("application/json"))
+            .Returns(jsonWriterMock.Object);
 
         requestAdapterMock
             .Setup(r => r.SerializationWriterFactory)
@@ -54,7 +79,7 @@ public class MultipartBodyTests
         writerMock.Setup(w => w.WriteStringValue("", "--" + body.Boundary));
         writerMock.Setup(w => w.WriteStringValue("Content-Type", "application/json"));
         writerMock.Setup(w => w.WriteStringValue("Content-Disposition", "form-data; name=\"file\""));
-        writerMock.Setup(w => w.WriteStringValue("", "fileContent"));
+        writerMock.Setup(w => w.WriteByteArrayValue("", ms.ToArray()));
         writerMock.Setup(w => w.WriteStringValue("", ""));
         writerMock.Setup(w => w.WriteStringValue("", "--" + body.Boundary + "--"));